move `sourceURL` tests to separate file
