+++ test/js/node/vm/vm.test.ts
@@ -687,68 +687,3 @@ resp.text().then((a) => {
     delete URL.prototype.ok;
   }
 });
-
-test("can get sourceURL from eval inside node:vm", () => {
-  try {
-    runInNewContext(
-      `
-throw new Error("hello");
-//# sourceURL=hellohello.js
-`,
-      {},
-    );
-  } catch (e: any) {
-    var err: Error = e;
-  }
-
-  expect(
-    err!
-      .stack!.replaceAll("\r\n", "\n")
-      .replaceAll(import.meta.path, "<this-url>")
-      .replaceAll(/<this-url>:\d+:\d+/g, "<this-url>:line:character"),
-  ).toMatchSnapshot();
-});
-
-test("can get sourceURL inside node:vm", () => {
-  const err = runInNewContext(
-    `
-
-function hello() {
-    return Bun.inspect(new Error("hello"));
-}
-
-hello();
-
-//# sourceURL=hellohello.js
-`,
-    { Bun },
-  );
-
-  expect(
-    err
-      .replaceAll("\r\n", "\n")
-      .replaceAll(import.meta.path, "<this-url>")
-      .replaceAll(/<this-url>:\d+:\d+/g, "<this-url>:line:character"),
-  ).toMatchSnapshot();
-});
-
-test("eval sourceURL is correct", () => {
-  const err = eval(
-    `
-
-function hello() {
-    return Bun.inspect(new Error("hello"));
-}
-
-hello();
-
-//# sourceURL=hellohello.js
-`,
-  );
-  expect(
-    err
-      .replaceAll("\r\n", "\n")
-      .replaceAll(import.meta.path, "<this-url>")
-      .replaceAll(/<this-url>:\d+:\d+/g, "<this-url>:line:character"),
-  ).toMatchSnapshot();
-});
